git-svn-id: https://subversion.imagemagick.org/subversion/ImageMagick/branches/ImageMagick-6@15835 aa41f4f7-0bf4-0310-aa73-e5a19afd5a74

+++ magick/quantize.c
@@ -3099,8 +3099,8 @@ static void ReduceImageColors(const Image *image,CubeInfo *cube_info)
             quantize_error);
           qsort(quantize_error,cube_info->nodes,sizeof(MagickRealType),
             MagickRealTypeCompare);
-          cube_info->next_threshold=quantize_error[MagickMax(cube_info->nodes-
-            110*(cube_info->maximum_colors+1)/100,0)];
+          cube_info->next_threshold=quantize_error[MagickMax((ssize_t)
+            cube_info->nodes-110*(cube_info->maximum_colors+1)/100,0)];
           quantize_error=(MagickRealType *) RelinquishMagickMemory(
             quantize_error);
         }
